@@ -1,9 +1,8 @@
-import { BrowserModule } from '@angular/platform-browser';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { MatButtonModule, MatCardModule, MatInputModule, MatListModule, MatToolbarModule, MatGridListModule, MatButtonToggleModule, MatRippleModule, MatSelectModule } from '@angular/material';
+import { MatButtonModule, MatCardModule, MatInputModule, MatListModule, MatToolbarModule, MatGridListModule, MatButtonToggleModule, MatRippleModule, MatSelectModule, MatTableModule, MatPaginator, MatPaginatorModule, MatProgressSpinnerModule } from '@angular/material';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { CReposTableService } from './shared/CReposTable/crepos-table.service';
@@ -11,7 +10,6 @@ import { CReposTableListComponent } from './crepos-table-list/crepos-table-list.
 import { CReposMatchSetListComponent } from './crepos-match-set-list/crepos-match-set-list.component';
 import { ScrollingModule } from '@angular/cdk/scrolling';
 import { CReposMatchRuleListComponent } from './crepos-match-rule-list/crepos-match-rule-list.component';
-import { DataTablesModule } from 'angular-datatables';
 import { CReposColumnListComponent } from './crepos-column-list/crepos-column-list.component';
 import { MatchReportComponent } from './match-report/match-report.component';
 import { HeaderComponent } from './header/header.component';
@@ -20,7 +18,6 @@ import { DashboardComponent } from './dashboard/dashboard.component';
 import { CreateReportComponent } from './create-report/create-report.component';
 import { MatchReportListComponent } from './match-report-list/match-report-list.component';
 import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { ChartsModule, WavesModule, MDBBootstrapModule } from 'angular-bootstrap-md';
 import { PieChartComponent } from './pie-chart/pie-chart.component';
 import { DoughnutChartComponent } from './doughnut-chart/doughnut-chart.component';
@@ -28,8 +25,10 @@ import { BarChartComponent } from './bar-chart/bar-chart.component';
 import { LoginComponent } from './login/login.component';
 import { JwtInterceptor } from './_service/jwt.interceptor';
 import { PackageViewerComponent } from './package-viewer/package-viewer.component';
-import { Ng4LoadingSpinnerModule } from 'ng4-loading-spinner';
 import { DatatableComponent } from './datatable/datatable.component';
+import { MatTableExporterModule } from 'mat-table-exporter';
+import { DataTablesModule } from 'angular-datatables';
+import { Ng4LoadingSpinnerModule } from 'ng4-loading-spinner';
 
 
 @NgModule({
@@ -53,7 +52,7 @@ import { DatatableComponent } from './datatable/datatable.component';
     DatatableComponent,
   ],
   imports: [
-    BrowserModule,
+    BrowserAnimationsModule,
     AppRoutingModule,
     HttpClientModule,
     BrowserAnimationsModule,
@@ -67,13 +66,15 @@ import { DatatableComponent } from './datatable/datatable.component';
     MatRippleModule,
     MatGridListModule,
     ScrollingModule,
-    DataTablesModule,
+    MatTableModule,
+    MatPaginatorModule,
+    MatProgressSpinnerModule,
     NgbModule,
-    FormsModule,
-    ReactiveFormsModule,
+    MatTableExporterModule,
+    Ng4LoadingSpinnerModule,
     ChartsModule,
+    DataTablesModule,
     WavesModule,
-    Ng4LoadingSpinnerModule.forRoot(),
     MDBBootstrapModule.forRoot()
   ],
   providers: [
