@@ -2,10 +2,10 @@ import { MMatchColumnComponent } from './mmatchcolumncomponent.model';
 
 export class MMatchColumn {
 
-    constructor(rowidMatchColumn: string) {
-        this.rowidMatchColumn = rowidMatchColumn;
+    constructor(rowidMmatchColumn: string) {
+        this.rowidMmatchColumn = rowidMmatchColumn;
     }
 
-    rowidMatchColumn: string;
+    rowidMmatchColumn: string;
     mmatchColumnComponents: MMatchColumnComponent[];
 }
